+++ src/assert_fields.rs
@@ -2,23 +2,44 @@
 ///
 /// # Examples
 ///
-/// This may be used when types have odd fields as a result of `#[cfg]`.
+/// One common use case is when types have fields defined multiple times as a
+/// result of `#[cfg]`. This can be an issue when exposing a public API.
 ///
-/// ```
+#[cfg_attr(feature = "nightly", doc = "```ignore")]
+#[cfg_attr(not(feature = "nightly"), doc = "```")]
 /// # #[macro_use] extern crate static_assertions;
-/// # fn main() {
-/// struct Ty {
+/// pub struct Ty {
 ///     #[cfg(windows)]
-///     value: u8,
+///     pub val1: u8,
 ///     #[cfg(not(windows))]
-///     value: usize,
+///     pub val1: usize,
+///
+///     #[cfg(unix)]
+///     pub val2: u32,
+///     #[cfg(not(unix))]
+///     pub val2: usize,
 /// }
 ///
-/// /* ... */
+/// // Requires a unique label in module scope
+/// assert_fields!(windows; Ty, val1);
 ///
-/// // Always have `value` regardless of OS
-/// assert_fields!(Ty, value);
-/// # }
+/// fn main() {
+///     // Always have `val2` regardless of OS
+///     assert_fields!(Ty, val2);
+/// }
+/// ```
+///
+/// The [labeling limitation](index.html#limitations) is not necessary if
+/// compiling on nightly Rust with the `nightly` feature enabled:
+///
+#[cfg_attr(feature = "nightly", doc = "```")]
+#[cfg_attr(not(feature = "nightly"), doc = "```ignore")]
+/// #![feature(underscore_const_names)]
+/// # #[macro_use] extern crate static_assertions;
+///
+/// use std::ops::Range;
+///
+/// assert_fields!(Range<u32>, start, end);
 /// ```
 ///
 /// Range does not have a field named `middle`:
@@ -26,13 +47,31 @@
 /// ```compile_fail
 /// # #[macro_use] extern crate static_assertions;
 /// # fn main() {
-/// use std::ops::Range;
-///
+/// # use std::ops::Range;
 /// assert_fields!(Range<u32>, middle);
 /// # }
 /// ```
 #[macro_export]
 macro_rules! assert_fields {
+    ($($xs:tt)+) => { _assert_fields!($($xs)+); };
+}
+
+#[doc(hidden)]
+#[cfg(feature = "nightly")]
+#[macro_export]
+macro_rules! _assert_fields {
+    ($t:path, $($f:ident),+) => {
+        #[allow(unknown_lints, unneeded_field_pattern)]
+        const _: fn() -> () = || {
+            $(let $t { $f: _, .. };)+
+        };
+    };
+}
+
+#[doc(hidden)]
+#[cfg(not(feature = "nightly"))]
+#[macro_export]
+macro_rules! _assert_fields {
     ($t:path, $($f:ident),+) => {
         #[allow(unknown_lints, unneeded_field_pattern)]
         { $(let $t { $f: _, .. };)+ }