+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog] and this project adheres to
 [Semantic Versioning].
 
 ## [Unreleased]
+### Removed
+- No need to use `macro!(label; ...)` syntax when compiling on nightly Rust and
+enabling the `nightly` feature flag.
 
 ## [0.2.5] - 2017-12-12
 ### Changed